Alternative class attribute implementation
What exactly changed from the user side:

    doc.add_class('highlight', 'today')  -> doc.classes.add('highlight today')
    doc.remove_class('news')             -> doc.classes.remove('news')
    doc.toggle_class('active')           -> doc.classes.toggle('active')

Main benefit from this change is ability to use full power of python's built-in
`set()` instead of restricted `add_class`, `remove_class`, `toggle_class` api.
Also, users are not forced to learn new apis it is enaugh to tell, that
`doc.classes` is a `set()` and everyone should know what to do with it.

Internally `class` attribute is always `set()` so it simplifies internal code
too. There is no need to convert string to set and set back to string for each
operation.