Add data attribute method to simpledoc
This commit adds a `simpledoc.data` method that adds HTML5 data attributes to the current tag. The method is called in the same way as `simpledoc.attr`. It takes keyword arguments or (key, value) pairs, prefixes each key with "data-", then updates the current tag with the new attributes.

This commit also includes documentation on the method.