Use 'phor' to give tags the 'for' attribute.
'for' is another reserved keyword like 'class', and will not work
as a keyword parameter for 'tag()', 'stag()', etc'. Allow 'phor'
to be passed to specify the 'for' attribute, instead. Additionally,
the new structure of the code will allow more substitutions to be
easily added by placing entries in the dict '_attr_key_translations'
in 'simpledoc.py'